+++ include/access/sequential.h
@@ -13,11 +13,11 @@ namespace Access {
         private:
             size_t block_size;
             size_t limit;
-            size_t current_offset;
+            off_t current_offset;
 
         public:
             SequentialAccess();
-            size_t nextOffset(void);
+            off_t nextOffset(void);
             void validate(void) const;
             friend void from_json(const json& j, SequentialAccess& config);
     };
